// src/pages/components/queries/queries.generated.ts
export type ProductCardFragment = { productId: string, title?: string | null, totalReviews?: number | null, ratingScore?: number | null, labels?: Array<string> | null, images?: { productImages?: Array<{ isMainImage?: boolean | null, status?: string | null, image?: { id: string, altText?: string | null, fileName: string, productCard?: string | null } | null }> | null } | null, location?: { country?: string | null, city?: string | null } | null, uri?: { country?: string | null, city?: string | null, url?: string | null } | null, priceFrom?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null, partner?: { id: string, name?: string | null, email?: string | null, website?: string | null, otherInfo?: string | null, salesEmail?: string | null, companyName?: string | null, companyPhone?: string | null, contactNumber?: string | null, accountsEmail?: string | null, contactPerson?: string | null, companyLogo?: { id: string, altText?: string | null, fileName: string, top3000?: string | null, top2400?: string | null, top1600?: string | null, top1400?: string | null, top1200?: string | null, top1140?: string | null, top1024?: string | null, top800?: string | null, top768?: string | null, top500?: string | null, top480?: string | null } | null } | null, duration?: { label?: string | null, maxDuration?: string | null, minDuration?: string | null } | null, legacy?: { shortId?: string | null, updatedAt?: any | null, forceAvailabilityV1?: boolean | null, subheader?: string | null, countDownLimit?: number | null, excludedBookingInfo?: string | null, bookingUrl?: string | null, openDatedBooking?: boolean | null, showMerchant?: boolean | null, hasNoIndexMetaTag?: boolean | null, hidePickupOnFrontend?: boolean | null, isDisplayGooglePay?: boolean | null, isDisplayPaypal?: boolean | null, isDisplayApplePay?: boolean | null, isDisplayAfterpay?: boolean | null, defaultDeparturePoint?: string | null, isDisplayHotelPickupWarning?: boolean | null, priceFrom?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null, displayPrice?: { adultPrice?: { type?: string | null, price?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null, value?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null } | null, childPrice?: { type?: string | null, price?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null, value?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null } | null, familyPrice?: { type?: string | null, price?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null, value?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null } | null } | null, recommendedRetailPrice?: { currencyCode: string, amount: number, convertedAmount?: number | null } | null } | null };